requirements.txt
Django>=3
djangorestframework
python-dotenv
json_log_formatter
python-logstash-async
